+++ src/apps/plutus/plutus.definition.ts
@@ -15,34 +15,48 @@ export const PLUTUS_DEFINITION = appDefinition({
       type: GroupType.TOKEN,
       label: 'plsDPX',
     },
+
     plsJones: {
       id: 'pls-jones',
       type: GroupType.TOKEN,
       label: 'plsJONES',
     },
+
     lock: {
       id: 'lock',
       type: GroupType.POSITION,
       label: 'Locked PLS',
       groupLabel: 'Farms',
     },
-    jones: {
+
+    plsJonesFarm: {
       id: 'jones',
       type: GroupType.POSITION,
-      label: 'Staked PlsJones',
-      groupLabel: 'Farms',
+      label: 'Staked plsJONES',
     },
-    dpx: {
-      id: 'dpx',
+
+    plsDpxFarm: {
+      id: 'pls-dpx-farm',
       type: GroupType.POSITION,
-      label: 'Staked PlsDPX',
-      groupLabel: 'Farms',
+      label: 'Staked plsDPX',
     },
-    stake: {
-      id: 'stake',
+
+    plsDpxFarmV2: {
+      id: 'pls-dpx-farm-v2',
+      type: GroupType.POSITION,
+      label: 'Staked plsDPX',
+    },
+
+    plsFarm: {
+      id: 'pls-farm',
       type: GroupType.POSITION,
       label: 'Staked PLS',
-      groupLabel: 'Farms',
+    },
+
+    tge: {
+      id: 'tge',
+      type: GroupType.POSITION,
+      label: 'Private TGE',
     },
   },
   tags: [AppTag.ASSET_MANAGEMENT, AppTag.FARMING],