+++ src/apps/stargate/avalanche/stargate.balance-fetcher.ts
@@ -53,7 +53,10 @@ export class AvalancheStargateBalanceFetcher implements BalanceFetcher {
       }),
       resolveClaimableTokenBalances: this.appToolkit.helpers.masterChefDefaultClaimableBalanceStrategy.build({
         resolveClaimableBalance: ({ multicall, contract, contractPosition, address }) =>
-          multicall.wrap(contract).pendingStargate(contractPosition.dataProps.poolIndex, address),
+          multicall
+            .wrap(contract)
+            .pendingStargate(contractPosition.dataProps.poolIndex, address)
+            .catch(() => 0),
       }),
     });
   }